svn merge -c 29071 https://source.sakaiproject.org/svn/chat/trunk U    chat-impl/impl/src/java/org/sakaiproject/chat2/model/impl/ChatManagerImpl.java svn log -r 29071 https://source.sakaiproject.org/svn/chat/trunk ------------------------------------------------------------------------ r29071 | chmaurer@iupui.edu | 2007-04-18 12:24:17 -0400 (Wed, 18 Apr 2007) | 2 lines
http://bugs.sakaiproject.org/jira/browse/SAK-9607
Some careless copy/paste.  I should have paid more attention to what the code was really doing!  I think it was closing the db session when clearing all settings for the context.  Then, when it tried to set the current room as the default, it had to create a new session.  Apparently lots of extra overhead!
------------------------------------------------------------------------


git-svn-id: https://source.sakaiproject.org/svn/chat/branches/sakai_2-4-x@29099 66ffb92e-73f9-0310-93c1-f5514f145a0a